+++ functions/randomHand.js
@@ -1,14 +1,15 @@
 const main = require('../index');
-const fetchDeck = require('./fetch').fetchDeck;
-const buildAttachment = require('./buildAttachment').buildAttachment;
-const _ = require('lodash');
+const {fetchDeck, getFlagNumber} = require('./fetch');
+const {buildAttachment} = require('./buildAttachment');
+const {shuffle, sortBy} = require('lodash');
 
 const randomHand = async (msg, params, flags) => {
 	const deck = await fetchDeck(params.join('+'));
+	const number = getFlagNumber(flags, 6);
 	let text, attachment;
 	if (deck) {
 		//grab 6 random cards
-		const randomCards = _.sortBy([...Array(6)].map(() => _.pullAt(deck.cards.sort(() => Math.random() - 0.5), _.random(deck.cards.length - 1))[0]), ['house', 'card_number']);
+		const randomCards = sortBy(shuffle(deck.cards).slice(0, Math.min(number ? number : 6, 8)), ['house', 'card_number']);
 
 		//build Title
 		const name = randomCards.map(card => `${card.card_number}`).join('_') + '.png';